Refactor : refactor codes (5)
