+++ src/main/java/http/request/QueryStrings.java
@@ -1,4 +1,4 @@
-package http.request.requestline;
+package http.request;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -6,9 +6,6 @@ import java.util.Map;
 
 public class QueryStrings {
     private Map<String, String> queryStrings = new HashMap<>();
-
-    public QueryStrings() {
-    }
 
     public QueryStrings(Map<String, String> queryStrings) {
         this.queryStrings = Collections.unmodifiableMap(queryStrings);
