+++ src/main/java/http/request/RequestLine.java
@@ -1,4 +1,4 @@
-package http.request.requestline;
+package http.request;
 
 import java.util.Collections;
 import java.util.Map;
