+++ src/main/java/http/response/ContentType.java
@@ -1,4 +1,4 @@
-package http.response.headers;
+package http.response;
 
 import java.util.Arrays;
