+++ src/main/java/http/response/HttpStatus.java
@@ -1,6 +1,4 @@
-package http.response.responseline;
-
-import java.util.Arrays;
+package http.response;
 
 public enum HttpStatus {
     OK(200),
