+++ src/main/java/http/response/ResponseBody.java
@@ -1,4 +1,4 @@
-package http.response.body;
+package http.response;
 
 public class ResponseBody {
     private byte[] body;
