+++ src/main/java/webserver/handler/custom/LoginHandler.java
@@ -2,12 +2,12 @@ package webserver.handler.custom;
 
 import db.DataBase;
 import http.request.Request;
-import http.request.headers.Headers;
-import http.request.requestline.QueryStrings;
-import http.response.headers.ContentType;
-import http.response.responseline.HttpStatus;
+import http.request.Headers;
+import http.request.QueryStrings;
+import http.response.ContentType;
+import http.response.HttpStatus;
 import http.response.Response;
-import http.response.body.ResponseBody;
+import http.response.ResponseBody;
 import model.User;
 import utils.FileIoUtils;
 import webserver.handler.Handler;
