+++ src/test/java/webserver/handler/custom/TemplateResourcesHandlerTest.java
@@ -1,16 +1,15 @@
 package webserver.handler.custom;
 
 import http.request.Request;
-import http.request.body.RequestBody;
-import http.request.headers.Headers;
-import http.request.requestline.RequestLine;
-import http.response.headers.ContentType;
-import http.response.responseline.HttpStatus;
+import http.request.RequestBody;
+import http.request.Headers;
+import http.request.RequestLine;
+import http.response.ContentType;
+import http.response.HttpStatus;
 import http.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import utils.FileIoUtils;
-import webserver.handler.custom.TemplateHandler;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
